\name{dgmisc-package}
\alias{dgmisc-package}
\alias{dgmisc}
\docType{package}
\title{Miscellaneous R functions for research}

\description{This package provides a vehicle for documenting, distributing and
maintaining otherwise disjointed bits of code used for research.
}

\details{
\tabular{ll}{
Package: \tab dgmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-13\cr
License: \tab GPL-2\cr
}

}
\author{
Code contributed by: Daniel Pritchard, Gretchen Brownstein

Package maintained by Daniel Pritchard (www.pritchard.co) 

Hosted on GitHub (https://github.com/dpritchard/dgmisc)
}
