\name{fr_functions}
\alias{fr_responses}
\alias{rogersII}
\alias{rogersII_fit}
\alias{rogersII_nll}
\alias{hollingsII}

\title{Support functions for fitting predator-prey functional responses.}
\description{DESCRIPTION}

\usage{
	fr_responses(show=FALSE)
	rogersII(X, a, h, P, T)
	rogersII_fit(data, samp, start, fixed, windows=FALSE)
	rogersII_nll(a, h, T, P, X, Y)
	hollingsII(X,a,h)
	}

\arguments{
  \item{X }{DESC.}
  \item{Y }{DESC.}
  \item{a, h, P, T }{DESC.}
  \item{data }{DESC.}
  \item{samp }{DESC.}
  \item{start }{DESC.}
  \item{fixed }{DESC.}
  \item{windows }{DESC.}
  \item{show }{DESC.}
}
\details{
TODO  
}

\author{
Daniel Pritchard
}

\seealso{
\code{\link{fr_fit}}.
}

\examples{
# TODO
}
